Remove duplicate streaming client creation code
