+++ agogosml/agogosml/common/abstract_streaming_client.py
@@ -2,7 +2,11 @@
 
 from abc import ABC
 from abc import abstractmethod
+from functools import lru_cache
 from typing import Dict
+from typing import Type
+
+from agogosml.utils.imports import find_implementations
 
 
 class AbstractStreamingClient(ABC):
@@ -29,3 +33,34 @@ class AbstractStreamingClient(ABC):
     def start_receiving(self, *args, **kwargs):
         """Start receiving messages from streaming client."""
         pass
+
+
+StreamingClientType = Type[AbstractStreamingClient]
+
+
+@lru_cache(maxsize=1)
+def find_streaming_clients() -> Dict[str, StreamingClientType]:
+    """
+    >>> senders = find_streaming_clients()
+    >>> sorted(senders.keys())
+    ['eventhub', 'kafka', 'mock']
+    """
+    return {
+        client.__name__.replace('StreamingClient', '').lower(): client
+        for client in find_implementations(AbstractStreamingClient)
+    }
+
+
+def create_streaming_client_from_config(config: dict) -> AbstractStreamingClient:
+    try:
+        client_config = config['client']['config']
+        client_type = config['client']['type']
+    except KeyError:
+        raise Exception('client cannot be empty')
+
+    try:
+        client_class = find_streaming_clients()[client_type]
+    except KeyError:
+        raise Exception('Unknown client type')
+
+    return client_class(client_config)
